Move linker component to separate react component
